pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = "a django blog"
authors = ["Adnene_wsl <adnen.guessoum@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-summernote = "^0.8.20.0"
pillow = "^9.3.0"
django-cors-headers = "^3.13.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^0.21.0"
flake8 = "^5.0.4"
pylint = "^2.15.5"
pipreqs = "^0.4.11"
black = "^22.10.0"
mypy = "0.982"
django-stubs = {extras = ["compatible-mypy"], version = "^1.13.0"}
pylint-django = "^2.5.3"
djangorestframework-stubs = {version = "1.2.0", extras = ["compatible-mypy"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# [tool.pylint]
# load-plugins = "pylint_django"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
mypy_path = ["./myblog", "."]

[tool.django-stubs]
django_settings_module = "myblog.settings"
